pyproject.toml
[project]
name = "langchain-chat-models-prompt-chains-rag-agents-tools"
version = "0.1.0"
description = ""
authors = [
    {name = "Sayan De",email = "sayandeten@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain (>=0.3.14,<0.4.0)",
    "langchain-openai (>=0.2.14,<0.3.0)",
    "langchain-ollama (>=0.2.2,<0.3.0)",
    "chromadb (>=0.6.3,<0.7.0)",
    "langchain-community (>=0.3.14,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
